Moved XCTestCase extension to separate file to avoid future redeclarations
