+++ PowerupUITests/XCTestCase+Wait.swift
@@ -1,0 +1,17 @@
+import XCTest
+
+// a manual wait function to be able to check that a popup is dismissed automatically
+extension XCTestCase {
+
+    func wait(for duration: TimeInterval) {
+        let waitExpectation = expectation(description: "Waiting")
+
+        let when = DispatchTime.now() + duration
+        DispatchQueue.main.asyncAfter(deadline: when) {
+            waitExpectation.fulfill()
+        }
+
+        waitForExpectations(timeout: duration)
+    }
+
+}
